requirements.txt
appdirs==1.4.4
attrs==21.2.0
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
desert==2020.11.18
flake8==3.9.2
idna==3.2
marshmallow==3.13.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
tomli==1.2.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
